Updating RoomsManager

+++ SchoolCommand/RoomManager.cs
@@ -10,9 +10,12 @@ namespace SchoolCommand
     {
         public static bool AddRoom(String name, String location)
         {
+            if (String.IsNullOrEmpty(name) || String.IsNullOrEmpty(location))
+                throw new InvalidOperationException("Cannot add new room with " + (String.IsNullOrEmpty(name) ? "name" : "location") + " empty.");
             var room = new Rooms
             {
-
+                Name = name,
+                Location = location
             };
 
             using (var db = new Entities())
@@ -54,8 +57,10 @@ namespace SchoolCommand
             using (var db = new Entities())
             {
                 var person = db.Rooms.Find(roomId);
-
-                db.Rooms.Remove(person);
+                if (person != null)
+                    db.Rooms.Remove(person);
+                else
+                    return false;
 
                 if (db.SaveChanges() > 1)
                     return true;
@@ -68,16 +73,36 @@ namespace SchoolCommand
         {
             using (var db = new Entities())
             {
-                //var select1 = (phone != null) ? db.People.Where(x => phone.Equals(phone)) : (age != null) ? db.People.Where(x=> x.Age.Equals(age)) : 
-                //    (address != null)? db.People.Where(x => x.Address.Equals(address)) : ((name != null)? db.People.Where( p => p.Name.Equals(name)) : null);
+                
+                String selectClause = "SELECT * FROM Rooms";
+                String whereClause = string.Empty;
+                String sql = string.Empty;
 
-                var select1 = from r in db.Rooms
-                              where ((name != null) ? r.Name == name : true) &&
-                              ((location != null) ? r.Location == location : true)
-                              select r;
-                return select1;
+                if (!string.IsNullOrEmpty(name))
+                {
+                    whereClause = "Rooms.Name LIKE '%" + name + "%'";
+                }
+                if (!string.IsNullOrEmpty(location))
+                {
+                    if (!string.IsNullOrEmpty(location))
+                        whereClause += " AND ";
+                    whereClause += "Rooms.Phone LIKE '%" + location + "%'";
+                }
 
+                if (!string.IsNullOrEmpty(whereClause))
+                {
+                    sql = selectClause + " WHERE " + whereClause;
+                }
+                else
+                {
+                    sql = selectClause;
+                }
 
+                var select1 = db.Rooms.SqlQuery(sql);
+
+                List<Rooms> results = select1.ToList<Rooms>();
+
+                return results;
             }
         }
     }
